Codigo atualizado com boas praticas de programação php de acordo com PSR2. Adicionado alguns comentarios. Usando metodo listar de questões.

+++ app/Controllers/salvarQuestao.php
@@ -1,11 +1,12 @@
 <?php
 require_once '../../vendor/autoload.php';
 use App\Models\Questao;
-if($_POST["metodo"]==0){
+if($_POST["metodo"]==0) {
 $_POST = json_decode(file_get_contents('php://input'), true);
 $tipo=$_POST["tipoQ"];
 $texto=$_POST["texto"];
 //trim e strip
+
 //carregando um aluno
 $q = new Questao();
 Questao::create(array(
@@ -13,7 +14,7 @@ Questao::create(array(
     'texto'=>$texto,
 ));
 }
-else{
+else if($_POST["metodo"]==1) {
     $valor=$_POST["idQ"];
     echo $valor;
     Questao::delete(array(
